Refactor Productos resource and fix user model relationship
Updated Productos resource to improve error handling, add duplicate name checks, and refactor CRUD methods to use producto_id as a parameter. Changed Usuario model's relationship attribute from 'compra' to 'compras' for consistency. These changes enhance API robustness and data integrity.

+++ backend/main/resources/productos.py
@@ -1,16 +1,10 @@
-from flask_restful import Resource, reqparse
+from flask_restful import Resource, reqparse, abort
 from flask import make_response, jsonify, request
 from .. import db
 from ..models import ProductoModel
 
 
 class Productos(Resource):
-    @staticmethod
-    def get():
-        productos = ProductoModel.query.all()
-        return make_response(
-            jsonify([producto.to_json() for producto in productos]), 200
-        )
 
     @staticmethod
     def post():
@@ -26,6 +20,13 @@ class Productos(Resource):
         parser.add_argument("stock", type=int, default=0)
         args = parser.parse_args()
 
+        # Check if a product with the same name already exists
+        existing_producto = ProductoModel.query.filter_by(nombre=args["nombre"]).first()
+        if existing_producto:
+            return make_response(
+                jsonify({"message": "Producto with this name already exists"}), 400
+            )
+
         new_producto = ProductoModel.from_json(args)
         db.session.add(new_producto)
         db.session.commit()
@@ -33,7 +34,17 @@ class Productos(Resource):
         return make_response(jsonify(new_producto.to_json()), 201)
 
     @staticmethod
-    def put():
+    def get():
+        productos = ProductoModel.query.all()
+        return make_response(
+            jsonify({
+                "productos": [producto.to_json() for producto in productos]
+                    }
+            ),
+        )
+
+    @staticmethod
+    def put(producto_id):
         parser = reqparse.RequestParser()
         parser.add_argument("id", type=int, required=True, help="ID is required")
         parser.add_argument("nombre", type=str, required=False)
@@ -43,7 +54,12 @@ class Productos(Resource):
         parser.add_argument("stock", type=int, required=False)
         args = parser.parse_args()
 
-        producto = ProductoModel.query.get_or_404(args["id"])
+        producto = ProductoModel.query.get(producto_id)
+
+        if not producto:
+            abort(404, message="Producto no encontrado")
+
+
         if args["nombre"] is not None:
             producto.nombre = args["nombre"]
         if args["descripcion"] is not None:
@@ -59,25 +75,33 @@ class Productos(Resource):
         return make_response(jsonify(producto.to_json()), 200)
 
     @staticmethod
-    def patch():
+    def patch(producto_id):
         parser = reqparse.RequestParser()
         parser.add_argument("id", type=int, required=True, help="ID is required")
         parser.add_argument("stock", type=int, required=True, help="Stock is required")
         args = parser.parse_args()
 
-        producto = ProductoModel.query.get_or_404(args["id"])
+        producto = ProductoModel.query.get(producto_id)
+
+        if not producto:
+            abort(404, message="Producto no encontrado")
+
         producto.stock = args["stock"]
 
         db.session.commit()
         return make_response(jsonify(producto.to_json()), 200)
 
     @staticmethod
-    def delete():
+    def delete(producto_id):
         parser = reqparse.RequestParser()
         parser.add_argument("id", type=int, required=True, help="ID is required")
         args = parser.parse_args()
 
-        producto = ProductoModel.query.get_or_404(args["id"])
+        producto = ProductoModel.query.get(producto_id)
+
+        if not producto:
+            abort(404, message="Producto no encontrado")
+
         db.session.delete(producto)
         db.session.commit()
         return make_response(jsonify({"message": "Producto deleted"}), 204)
@@ -86,12 +110,20 @@ class Productos(Resource):
 class ProductoDetail(Resource):
     @staticmethod
     def get(producto_id):
-        producto = ProductoModel.query.get_or_404(producto_id)
+
+        producto = ProductoModel.query.get(producto_id)
+        if not producto:
+            abort(404, message="Producto no encontrado")
+
         return make_response(jsonify(producto.to_json()), 200)
 
     @staticmethod
     def delete(producto_id):
-        producto = ProductoModel.query.get_or_404(producto_id)
+        producto = ProductoModel.query.get(producto_id)
+
+        if not producto:
+            abort(404, message="Producto no encontrado")
+            
         db.session.delete(producto)
         db.session.commit()
         return make_response(jsonify({"message": "Producto deleted"}), 204)
